Remove kotlinx.datetime.Instant and kotlinx.datetime.Clock
