+++ core/common/src/TimeZone.kt
@@ -179,24 +179,6 @@ public expect open class TimeZone {
     public fun Instant.toLocalDateTime(): LocalDateTime
 
     /**
-     * Return the civil datetime value that this instant has in the time zone provided as an implicit receiver.
-     *
-     * Note that while this conversion is unambiguous, the inverse ([LocalDateTime.toInstant])
-     * is not necessarily so.
-     *
-     * @see LocalDateTime.toInstant
-     * @see Instant.offsetIn
-     * @throws DateTimeArithmeticException if this value is too large to fit in [LocalDateTime].
-     * @sample kotlinx.datetime.test.samples.TimeZoneSamples.toLocalDateTimeWithTwoReceivers
-     */
-    @Suppress("DEPRECATION")
-    @Deprecated("kotlinx.datetime.Instant is superseded by kotlin.time.Instant",
-        level = DeprecationLevel.WARNING,
-        replaceWith = ReplaceWith("this.toStdlibInstant().toLocalDateTime()")
-    )
-    public fun kotlinx.datetime.Instant.toLocalDateTime(): LocalDateTime
-
-    /**
      * Returns an instant that corresponds to this civil datetime value in the time zone provided as an implicit receiver.
      *
      * Note that the conversion is not always well-defined. There can be the following possible situations:
@@ -214,10 +196,6 @@ public expect open class TimeZone {
     @Suppress("DEPRECATION_ERROR")
     public fun LocalDateTime.toInstant(youShallNotPass: OverloadMarker = OverloadMarker.INSTANCE): Instant
 
-    @PublishedApi
-    @Suppress("INVISIBLE_MEMBER", "INVISIBLE_REFERENCE", "DEPRECATION")
-    @kotlin.internal.LowPriorityInOverloadResolution
-    internal fun LocalDateTime.toInstant(): kotlinx.datetime.Instant
 }
 
 /**
@@ -271,14 +249,6 @@ public typealias ZoneOffset = FixedOffsetTimeZone
  */
 public expect fun TimeZone.offsetAt(instant: Instant): UtcOffset
 
-@Suppress("DEPRECATION")
-@Deprecated("kotlinx.datetime.Instant is superseded by kotlin.time.Instant",
-    level = DeprecationLevel.WARNING,
-    replaceWith = ReplaceWith("this.offsetAt(instant.toStdlibInstant())")
-)
-public fun TimeZone.offsetAt(instant: kotlinx.datetime.Instant): UtcOffset =
-    offsetAt(instant.toStdlibInstant())
-
 /**
  * Returns a civil datetime value that this instant has in the specified [timeZone].
  *
@@ -292,14 +262,6 @@ public fun TimeZone.offsetAt(instant: kotlinx.datetime.Instant): UtcOffset =
  */
 public expect fun Instant.toLocalDateTime(timeZone: TimeZone): LocalDateTime
 
-@Suppress("DEPRECATION")
-@Deprecated("kotlinx.datetime.Instant is superseded by kotlin.time.Instant",
-    level = DeprecationLevel.WARNING,
-    replaceWith = ReplaceWith("this.toStdlibInstant().toLocalDateTime(timeZone)")
-)
-public fun kotlinx.datetime.Instant.toLocalDateTime(timeZone: TimeZone): LocalDateTime =
-    toStdlibInstant().toLocalDateTime(timeZone)
-
 /**
  * Returns a civil datetime value that this instant has in the specified [UTC offset][offset].
  *
@@ -312,14 +274,6 @@ public fun kotlinx.datetime.Instant.toLocalDateTime(timeZone: TimeZone): LocalDa
  * @sample kotlinx.datetime.test.samples.TimeZoneSamples.instantToLocalDateTimeInOffset
  */
 internal expect fun Instant.toLocalDateTime(offset: UtcOffset): LocalDateTime
-
-@Suppress("DEPRECATION")
-@Deprecated("kotlinx.datetime.Instant is superseded by kotlin.time.Instant",
-    level = DeprecationLevel.WARNING,
-    replaceWith = ReplaceWith("this.toStdlibInstant().toLocalDateTime(offset)")
-)
-public fun kotlinx.datetime.Instant.toLocalDateTime(offset: UtcOffset): LocalDateTime =
-    toStdlibInstant().toLocalDateTime(offset)
 
 /**
  * Finds the offset from UTC the specified [timeZone] has at this instant of physical time.
@@ -335,13 +289,6 @@ public fun kotlinx.datetime.Instant.toLocalDateTime(offset: UtcOffset): LocalDat
 public fun Instant.offsetIn(timeZone: TimeZone): UtcOffset =
         timeZone.offsetAt(this)
 
-@Suppress("DEPRECATION")
-@Deprecated("kotlinx.datetime.Instant is superseded by kotlin.time.Instant",
-    level = DeprecationLevel.WARNING,
-    replaceWith = ReplaceWith("this.toStdlibInstant().offsetIn(timeZone)")
-)
-public fun kotlinx.datetime.Instant.offsetIn(timeZone: TimeZone): UtcOffset =
-    timeZone.offsetAt(toStdlibInstant())
 
 /**
  * Returns an instant that corresponds to this civil datetime value in the specified [timeZone].
@@ -361,12 +308,6 @@ public fun kotlinx.datetime.Instant.offsetIn(timeZone: TimeZone): UtcOffset =
 @Suppress("DEPRECATION_ERROR")
 public expect fun LocalDateTime.toInstant(timeZone: TimeZone, youShallNotPass: OverloadMarker = OverloadMarker.INSTANCE): Instant
 
-@PublishedApi
-@Suppress("INVISIBLE_MEMBER", "INVISIBLE_REFERENCE", "DEPRECATION")
-@kotlin.internal.LowPriorityInOverloadResolution
-internal fun LocalDateTime.toInstant(timeZone: TimeZone): kotlinx.datetime.Instant =
-    toInstant(timeZone).toDeprecatedInstant()
-
 /**
  * Returns an instant that corresponds to this civil datetime value that happens at the specified [UTC offset][offset].
  *
@@ -375,12 +316,6 @@ internal fun LocalDateTime.toInstant(timeZone: TimeZone): kotlinx.datetime.Insta
  */
 @Suppress("DEPRECATION_ERROR")
 public expect fun LocalDateTime.toInstant(offset: UtcOffset, youShallNotPass: OverloadMarker = OverloadMarker.INSTANCE): Instant
-
-@PublishedApi
-@Suppress("INVISIBLE_MEMBER", "INVISIBLE_REFERENCE", "DEPRECATION")
-@kotlin.internal.LowPriorityInOverloadResolution
-internal fun LocalDateTime.toInstant(offset: UtcOffset): kotlinx.datetime.Instant =
-    toInstant(offset).toDeprecatedInstant()
 
 /**
  * Returns an instant that corresponds to the start of this date in the specified [timeZone].
@@ -397,12 +332,6 @@ internal fun LocalDateTime.toInstant(offset: UtcOffset): kotlinx.datetime.Instan
  */
 @Suppress("DEPRECATION_ERROR")
 public expect fun LocalDate.atStartOfDayIn(timeZone: TimeZone, youShallNotPass: OverloadMarker = OverloadMarker.INSTANCE): Instant
-
-@PublishedApi
-@Suppress("INVISIBLE_MEMBER", "INVISIBLE_REFERENCE", "DEPRECATION")
-@kotlin.internal.LowPriorityInOverloadResolution
-internal fun LocalDate.atStartOfDayIn(timeZone: TimeZone): kotlinx.datetime.Instant =
-    atStartOfDayIn(timeZone).toDeprecatedInstant()
 
 internal expect fun localDateTimeToInstant(
     dateTime: LocalDateTime, timeZone: TimeZone, preferred: UtcOffset? = null
