+++ integration-testing/serialization/common/test/IntegrationTest.kt
@@ -21,7 +21,6 @@ class IntegrationTest {
 
     @Serializable
     data class Dummy(
-        @Contextual val instant: Instant,
         @Contextual val date: LocalDate,
         @Contextual val dateTime: LocalDateTime,
         @Contextual val datePeriod: DatePeriod,
@@ -31,7 +30,6 @@ class IntegrationTest {
         )
 
     private val module = SerializersModule {
-        contextual(InstantComponentSerializer)
         contextual(LocalDateComponentSerializer)
         contextual(LocalDateTimeComponentSerializer)
         contextual(DatePeriodComponentSerializer)
@@ -45,7 +43,6 @@ class IntegrationTest {
     @Test
     fun testContextualSerialization() {
         val dummyValue = Dummy(
-            Instant.parse("2021-03-24T01:29:30.123456789Z"),
             LocalDate.parse("2020-01-02"),
             LocalDateTime.parse("2020-01-03T12:59:58.010203045"),
             DatePeriod.parse("P20Y-2M-3D"),
@@ -53,8 +50,7 @@ class IntegrationTest {
             // DayOfWeek.MONDAY, // doesn't compile on Native
             // Month.DECEMBER,
         )
-        val json = """{"instant":{"epochSeconds":1616549370,"nanosecondsOfSecond":123456789},""" +
-                """"date":{"year":2020,"month":1,"day":2},""" +
+        val json = """{"date":{"year":2020,"month":1,"day":2},""" +
                 """"dateTime":{"year":2020,"month":1,"day":3,"hour":12,"minute":59,"second":58,"nanosecond":10203045},""" +
                 """"datePeriod":{"years":19,"months":10,"days":-3},""" +
                 """"dateTimePeriod":{"years":-49,"months":-11,"days":2,"hours":-3,"minutes":-4,"seconds":-5,"nanoseconds":-12300000}}"""
@@ -64,7 +60,6 @@ class IntegrationTest {
 
     @Serializable
     data class Dummy2(
-        val instant: Instant,
         val date: LocalDate,
         val dateTime: LocalDateTime,
         val datePeriod: DatePeriod,
@@ -76,7 +71,6 @@ class IntegrationTest {
     @Test
     fun testDefaultSerialization() {
         val dummyValue = Dummy2(
-            Instant.parse("2021-03-24T01:29:30.123456789Z"),
             LocalDate.parse("2020-01-02"),
             LocalDateTime.parse("2020-01-03T12:59:58.010203045"),
             DatePeriod.parse("P20Y-2M-3D"),
@@ -84,8 +78,7 @@ class IntegrationTest {
             // DayOfWeek.MONDAY,
             // Month.DECEMBER,
         )
-        val json = "{\"instant\":\"2021-03-24T01:29:30.123456789Z\"," +
-                "\"date\":\"2020-01-02\"," +
+        val json = "{\"date\":\"2020-01-02\"," +
                 "\"dateTime\":\"2020-01-03T12:59:58.010203045\"," +
                 "\"datePeriod\":\"P19Y10M-3D\"," +
                 "\"dateTimePeriod\":\"P-49Y-11M2DT-3H-4M-5.012300000S\"" +
@@ -96,7 +89,6 @@ class IntegrationTest {
 
     @Serializable
     data class Dummy3(
-        @Serializable(with = InstantComponentSerializer::class) val instant: Instant,
         @Serializable(with = LocalDateComponentSerializer::class) val date: LocalDate,
         @Serializable(with = LocalDateTimeComponentSerializer::class) val dateTime: LocalDateTime,
         @Serializable(with = DatePeriodComponentSerializer::class) val datePeriod: DatePeriod,
@@ -108,7 +100,6 @@ class IntegrationTest {
     @Test
     fun testExplicitSerializerSpecification() {
         val dummyValue = Dummy3(
-            Instant.parse("2021-03-24T01:29:30.123456789Z"),
             LocalDate.parse("2020-01-02"),
             LocalDateTime.parse("2020-01-03T12:59:58.010203045"),
             DatePeriod.parse("P20Y-2M-3D"),
@@ -116,8 +107,7 @@ class IntegrationTest {
             DayOfWeek.MONDAY,
             Month.DECEMBER,
         )
-        val json = """{"instant":{"epochSeconds":1616549370,"nanosecondsOfSecond":123456789},""" +
-                """"date":{"year":2020,"month":1,"day":2},""" +
+        val json = """{"date":{"year":2020,"month":1,"day":2},""" +
                 """"dateTime":{"year":2020,"month":1,"day":3,"hour":12,"minute":59,"second":58,"nanosecond":10203045},""" +
                 """"datePeriod":{"years":19,"months":10,"days":-3},""" +
                 """"dateTimePeriod":{"years":-49,"months":-11,"days":2,"hours":-3,"minutes":-4,"seconds":-5,"nanoseconds":-12300000},""" +
